mi_api_comentado/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
dnspython==2.4.2
pyrebase4==4.5.0
pyjwt==2.6.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==1.10.7
email-validator==2.0.0
python-multipart==0.0.6
